@@ -1,5 +1,5 @@
-class ActsAsTaggableOnMigration < ActiveRecord::Migration
-  def self.up
+class ActsAsTaggableOnMigration < ActiveRecord::Migration[5.0]
+  def up
     create_table :tags do |t|
       t.string :name
     end
@@ -23,7 +23,7 @@ class ActsAsTaggableOnMigration < ActiveRecord::Migration
     add_index :taggings, [:taggable_id, :taggable_type, :context]
   end
 
-  def self.down
+  def down
     drop_table :taggings
     drop_table :tags
   end
