@@ -1,16 +1,4 @@
-module Item
-  def images_count
-    count = 0
-    count += 1 if self.image_filename_1.present?
-    count += 1 if self.image_filename_2.present?
-    count += 1 if self.image_filename_3.present?
-    return count
-  end
-
-  def move
-    raise NotImplementedError
-  end
-
+module TaggableItem
   module ClassMethods
     def active_taggings
       ActsAsTaggableOn::Tagging.where(taggable_type: self.name, context: 'categories')
