@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -16,70 +15,67 @@ ActiveRecord::Schema.define(version: 20160408222110) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "admin_users", force: true do |t|
-    t.string   "email",                  default: "", null: false
-    t.string   "encrypted_password",     default: "", null: false
-    t.string   "reset_password_token"
+  create_table "admin_users", id: :serial, force: :cascade do |t|
+    t.string "email", default: "", null: false
+    t.string "encrypted_password", default: "", null: false
+    t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          default: 0
+    t.integer "sign_in_count", default: 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
-    t.string   "current_sign_in_ip"
-    t.string   "last_sign_in_ip"
-    t.datetime "created_at",                          null: false
-    t.datetime "updated_at",                          null: false
+    t.string "current_sign_in_ip"
+    t.string "last_sign_in_ip"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["email"], name: "index_admin_users_on_email", unique: true
+    t.index ["reset_password_token"], name: "index_admin_users_on_reset_password_token", unique: true
   end
 
-  add_index "admin_users", ["email"], name: "index_admin_users_on_email", unique: true, using: :btree
-  add_index "admin_users", ["reset_password_token"], name: "index_admin_users_on_reset_password_token", unique: true, using: :btree
-
-  create_table "portfolio_items", force: true do |t|
-    t.string  "title"
-    t.text    "description"
-    t.string  "image_filename_1"
-    t.string  "image_filename_2"
-    t.string  "image_filename_3"
-    t.string  "thumbnail_filename"
-    t.boolean "enabled",            default: true
+  create_table "portfolio_items", id: :serial, force: :cascade do |t|
+    t.string "title"
+    t.text "description"
+    t.string "image_filename_1"
+    t.string "image_filename_2"
+    t.string "image_filename_3"
+    t.string "thumbnail_filename"
+    t.boolean "enabled", default: true
   end
 
-  create_table "shop_items", force: true do |t|
-    t.string  "title"
-    t.text    "description"
-    t.string  "image_filename_1"
-    t.string  "image_filename_2"
-    t.string  "image_filename_3"
-    t.string  "thumbnail_filename"
-    t.boolean "enabled",                                    default: true
-    t.integer "quantity",                                   default: 1
-    t.decimal "price",              precision: 8, scale: 2, default: 0.0
+  create_table "shop_items", id: :serial, force: :cascade do |t|
+    t.string "title"
+    t.text "description"
+    t.string "image_filename_1"
+    t.string "image_filename_2"
+    t.string "image_filename_3"
+    t.string "thumbnail_filename"
+    t.boolean "enabled", default: true
+    t.integer "quantity", default: 1
+    t.decimal "price", precision: 8, scale: 2, default: "0.0"
   end
 
-  create_table "static_contents", force: true do |t|
+  create_table "static_contents", id: :serial, force: :cascade do |t|
     t.string "page"
     t.string "part"
-    t.text   "body"
+    t.text "body"
   end
 
-  create_table "taggings", force: true do |t|
-    t.integer  "tag_id"
-    t.integer  "taggable_id"
-    t.string   "taggable_type"
-    t.integer  "tagger_id"
-    t.string   "tagger_type"
-    t.string   "context",       limit: 128
+  create_table "taggings", id: :serial, force: :cascade do |t|
+    t.integer "tag_id"
+    t.integer "taggable_id"
+    t.string "taggable_type"
+    t.integer "tagger_id"
+    t.string "tagger_type"
+    t.string "context", limit: 128
     t.datetime "created_at"
+    t.index ["tag_id", "taggable_id", "taggable_type", "context", "tagger_id", "tagger_type"], name: "taggings_idx", unique: true
+    t.index ["taggable_id", "taggable_type", "context"], name: "index_taggings_on_taggable_id_and_taggable_type_and_context"
   end
 
-  add_index "taggings", ["tag_id", "taggable_id", "taggable_type", "context", "tagger_id", "tagger_type"], name: "taggings_idx", unique: true, using: :btree
-  add_index "taggings", ["taggable_id", "taggable_type", "context"], name: "index_taggings_on_taggable_id_and_taggable_type_and_context", using: :btree
-
-  create_table "tags", force: true do |t|
-    t.string  "name"
+  create_table "tags", id: :serial, force: :cascade do |t|
+    t.string "name"
     t.integer "taggings_count", default: 0
+    t.index ["name"], name: "index_tags_on_name", unique: true
   end
-
-  add_index "tags", ["name"], name: "index_tags_on_name", unique: true, using: :btree
 
 end
